tutor 26

+++ IG-4/technique.cpp
@@ -157,7 +157,8 @@ GLint Technique::GetUniformLocation(const char* pUniformName)
 {
     GLint Location = glGetUniformLocation(m_shaderProg, pUniformName);
 
-    if (Location == (GLint)0xFFFFFFFF) {
+    if (Location == 0xFFFFFFFF)
+    {
         fprintf(stderr, "Warning! Unable to get the location of uniform '%s'\n", pUniformName);
     }
